Refactor: ConnectionManager for WebSocket handling and update API endpoints
